+++ server/src/main/java/com/example/tyfserver/auth/config/OAuth2Config.java
@@ -1,8 +1,8 @@
-package com.example.tyfserver.oauth2.config;
+package com.example.tyfserver.auth.config;
 
-import com.example.tyfserver.oauth2.domain.GoogleOAuth2;
-import com.example.tyfserver.oauth2.domain.KakaoOAuth2;
-import com.example.tyfserver.oauth2.domain.NaverOAuth2;
+import com.example.tyfserver.auth.domain.GoogleOAuth2;
+import com.example.tyfserver.auth.domain.KakaoOAuth2;
+import com.example.tyfserver.auth.domain.NaverOAuth2;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Configuration;
