+++ server/src/main/java/com/example/tyfserver/auth/domain/OAuth2Type.java
@@ -1,4 +1,4 @@
-package com.example.tyfserver.oauth2.domain;
+package com.example.tyfserver.auth.domain;
 
 import org.json.JSONObject;
